% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesets2ids.R
\name{genesets2ids}
\alias{genesets2ids}
\title{Creates a binary matrix from a list of genes}
\usage{
genesets2ids(y, genesetlist)
}
\arguments{
\item{y}{numeric matrix. The gene expression matrix from Highly variable genes}

\item{genesetlist}{list of genesets. Each element of the list is a geneset.
Each geneset can contain genes or gene ids. However, the genes in the genesets must match the rownames in \code{y}.}
}
\value{
numeric matrix. A binary matrix is returned where rows are genes, columns are genesets.
}
\description{
Creates a binary matrix from a list of genes
}
