% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneDrug2CellScores.R
\name{pruneDrug2CellScores.fitLambda}
\alias{pruneDrug2CellScores.fitLambda}
\title{Fit multiple shrinkage values for Penalized generalized linear modelling for selection of relevant pre-computed Drug2Cell scores}
\usage{
pruneDrug2CellScores.fitLambda(
  data,
  embedding,
  suppress_plot = FALSE,
  lambda = "lambda.1se",
  gamma = 0,
  nfolds = 10,
  trace.it = 0
)
}
\arguments{
\item{data}{numeric matrix of Drug2Cell scores. columns are Drug2Cell scores, rows are cells or samples.}

\item{embedding}{numeric matrix. UMAP, PHATE or other cell embedding that captures transcriptional similarity of the cells.}

\item{suppress_plot}{logical. Should the plot of glm cross-validation results be suppressed? default to FALSE.}

\item{lambda}{numeric or character. The lasso penalty. This is set to optimal value selected
by cross-validation by default. Should be set to exp(log(lambda)), where
log(lambda) is chosen based on the diagnostic plot returned by the function
(requires \code{suppress_plot = FALSE}).}

\item{gamma}{numeric. Controls the behavior of shrinkage operator e.g. \code{gamma=0.5}
is equivalent to the Elastic Net model}

\item{nfolds}{numeric. Number of folds for cross-validation.}
}
\value{
a character vector of selected genesets. Regression coefficients from the glmnet model
can be accessed via the \code{glmnet_coef} attribute.
}
\description{
Fit multiple shrinkage values for Penalized generalized linear modelling for selection of relevant pre-computed Drug2Cell scores
}
\details{
This function should be used to determine the optimal shrinkage value \code{lambda} for geneset selection. 
This function is applicable to any pre-computed gene set and signature scores.
}
\author{
Soroor Hediyeh-zadeh
}
