% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruneGenesets.R
\name{pruneGenesets}
\alias{pruneGenesets}
\title{Penalized generalized linear models for model-based selection of relevant genesets}
\usage{
pruneGenesets(
  data,
  genesetlist,
  embedding,
  hvg = NULL,
  min_gs_size = 3,
  suppress_plot = FALSE,
  lambda = "lambda.1se",
  gamma = 0,
  nfolds = 10
)
}
\arguments{
\item{data}{numeric matrix of log-normalised values. Rows are genes, columns are cells or samples}

\item{genesetlist}{list of genesets. Each element of the list is a geneset.
Each geneset can contain genes or gene ids. However,
the genes in the genesets must match the rownames in \code{data}.}

\item{embedding}{numeric matrix. UMAP, tSNE or other cell embedding that captures transcriptional similarity of the cells.}

\item{hvg}{character vector of Highly Variable Genes.}

\item{min_gs_size}{numeric. Minimum number of genes per geneset.}

\item{suppress_plot}{logical. Should the plot of glm cross-validation results be suppressed? default to FALSE.}

\item{lambda}{numeric or character. The lasso penalty. This is set to optimal value selected
by cross-validation by default. Should be set to exp(log(lambda)), where
log(lambda) is chosen based on the diagnostic plot returned by the function
(requires \code{suppress_plot = FALSE}).}

\item{gamma}{numeric. Controls the behavior of shrinkage operator e.g. \code{gamma=0.5}
is equivalent to the Elastic Net model}

\item{nfolds}{numeric. Number of folds for cross-validation.}
}
\value{
a character vector of selected genesets. Regression coefficients from the glmnet model
can be accessed via the \code{glmnet_coef} attribute.
}
\description{
Penalized generalized linear models for model-based selection of relevant genesets
}
