% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDECAF.R
\name{scDECAF}
\alias{scDECAF}
\title{Reference-free cell type and phenotype annotation by learning geneset representations}
\usage{
scDECAF(
  data,
  gs,
  hvg,
  cca.k = NULL,
  standardize = TRUE,
  k = 30,
  thresh = 0.5,
  embedding,
  n_components = ncol(gs) - 1,
  max_iter = 1
)
}
\arguments{
\item{data}{numeric matrix of log-normalised values. Rows are genes, columns are cells or samples}

\item{gs}{binary matrix created by \code{genesets2ids}. Rows are genes, columns are genesets. Each column represents a geneset.}

\item{hvg}{character vector of Highly Variable Genes.}

\item{cca.k}{numeric. Number of canonical factors used to learn geneset representations.}

\item{standardize}{logical. Should input expression data be standardized? default to TRUE.}

\item{k}{numeric. Number of nearest neighbors used to refine cell-label assignment. Default to 30.}

\item{thresh}{numeric. Assignment confidence probability threshold. Default to 0.5.}

\item{embedding}{numeric matrix. UMAP, tSNE or other cell embedding that captures transcriptional similarity of the cells.}

\item{n_components}{numeric. Number of canonical factors to use. Default to all.}

\item{max_iter}{numeric. currently not used.}
}
\value{
Data frame of annotation results
}
\description{
Reference-free cell type and phenotype annotation by learning geneset representations
}
\details{
The cell-geneset assignments are determined by cosine similarity of vector representations
of cells and genesets by CCA. There is the a “re-assignment” step where the umap (embedding) coordinates of the cells are
used to refine labels based on a local (\code{k}=3-15) or global (\code{k}>15) neighborhood of
the cell, as the lower-dimensional embedding of the cells captures the transcriptional heterogeneity of
cells, and cells with similar transcriptional profiles should ideally have same labels/phenotypes.
Users have their own choice of embedding, as there are various methods to obtain low-dimensional embedding
for the cells, and each offer unique properties in terms of preservation of local or global structures, or transcriptional
heterogeneity. The input gene expression matrix is preferred to contain filtered, log-normalised values.
However, the algorithm is expected to work with RPKM and/or log-RPKM values as well.
}
\author{
Soroor Hediyeh-zadeh
}
