% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scDECAF.R
\name{scDECAF}
\alias{scDECAF}
\title{Reference-free cell type and phenotype annotation by learning geneset representations}
\usage{
scDECAF(data, gs, hvg, cca.k = NULL, standardize = TRUE, k = 30,
  thresh = 0.5, embedding, n_components = ncol(gs) - 1, max_iter = 1)
}
\arguments{
\item{data}{numeric matrix of log-normalised values. Rows are genes, columns are cells or samples}

\item{gs}{binary matrix created by \code{genesets2ids}. Rows are genes, columns are genesets. Each column represents a geneset.}

\item{hvg}{character vector of Highly Variable Genes.}

\item{cca.k}{numeric. Number of canonical factors used to learn geneset representations.}

\item{standardize}{logical. Should input expression data be standardized? default to TRUE.}

\item{k}{numeric. Number of nearest neighbors used to refine cell-label assignment. Default to 30.}

\item{thresh}{numeric. Assignment confidence probability threshold. Default to 0.5.}

\item{embedding}{numeric matrix. UMAP, tSNE or other cell embedding that captures transcriptional similarity of the cells.}

\item{n_components}{numeric. Number of canonical factors to use. Default to all.}

\item{max_iter}{numeric. currently not used.}
}
\value{
Data frame of annotation results. The \code{pred_celltype} and \code{reassigned_celltype} columns report initial and refined
label/geneset assignments respectively. The \code{score} is the cosine distance between the vector representation of the cell and the
closest geneset. The variable \code{uncertainty} is the confidence in label assignment based on the label of the cells in the neighborhood.
\code{prop_neighbourhood_preserved} is the proportion of nearest neighbors in embedding space preserved in CCA subspaces.
\code{dist_to_centroid} is the distance of the cell from label centroid, where label centroid is computed as the median coordinate
values for all cells assigned to that label in the embedding space. Raw cosine similarity scores for all genesets in the input
genesets lists, \code{gs}, can be accessed via the \code{raw_scores} attribute assigned to the output data-frame
i.e. \code{attributes(res_df)$raw_scores}, where \code{res_df} is the data-frame returned by the function.
}
\description{
Reference-free cell type and phenotype annotation by learning geneset representations
}
\details{
The cell-geneset assignments are determined by cosine similarity of vector representations
of cells and genesets by CCA. There is the a “re-assignment” step where the umap (embedding) coordinates of the cells are
used to refine labels based on a local (\code{k}=3-15) or global (\code{k}>15) neighborhood of
the cell, as the lower-dimensional embedding of the cells captures the transcriptional heterogeneity of
cells, and cells with similar transcriptional profiles should ideally have same labels/phenotypes.
Users have their own choice of embedding, as there are various methods to obtain low-dimensional embedding
for the cells, and each offer unique properties in terms of preservation of local or global structures, or transcriptional
heterogeneity. The input gene expression matrix is preferred to contain filtered, log-normalised values.
However, the algorithm is expected to work with RPKM and/or log-RPKM values as well.
}
\author{
Soroor Hediyeh-zadeh
}
